creating toymodel that reflects the overall multhithreaded structure of the SVD application.
No work is done in the GUI thread:
* one thread for the "core" model
* one thread for the "inference / dnn)

in the core model pixels are processed in parallel, and moved to a
"batch" structure. When a batch is full, the batch is forwarded to the
dnn-thread and processed. When finished in the DNN thread, the results
are incorporated in the core thread.
TODO: use parallel computation *within* the dnn thread (inference,
postprocessing)