Change DNN model interaction
Large change in architecture that allows multiple types of batches (e.g.
a DNN batch and a C++ batch) and moves some of the logic (selecting
states) closer to the main model.
Batch is a base class and derived variants (e.g. BatchDNN) add more
functionality. A batch has a type that is used to route the batch across
threads.
Needs still some testing.