Multiple DNN instances in SVD
Refactoring of the DNN code that allows to run multiple instances of a
DNN in parallel. Each DNN instance sets up the same network and
settings. On execution of a batch different DNN instances are selected
(cycling through all available DNNs).
Fixed a crash when stats are collected before BatchManager is available.


Former-commit-id: 6943e4eff8d16fe442d8d21e02af3703099ff142